[Rename]shareinventory -> sharedinventory
