+++ src/main/scala/com/github/unchama/seichiassist/subsystems/sharedinventory/SharedInventoryAPI.scala
@@ -1,9 +1,9 @@
-package com.github.unchama.seichiassist.subsystems.shareinventory
+package com.github.unchama.seichiassist.subsystems.sharedinventory
 
 import cats.effect.concurrent.Ref
 import com.github.unchama.datarepository.KeyedDataRepository
-import com.github.unchama.seichiassist.subsystems.shareinventory.domain.SharedFlag
-import com.github.unchama.seichiassist.subsystems.shareinventory.domain.bukkit.InventoryContents
+import com.github.unchama.seichiassist.subsystems.sharedinventory.domain.SharedFlag
+import com.github.unchama.seichiassist.subsystems.sharedinventory.domain.bukkit.InventoryContents
 
 import java.util.UUID
