+++ src/main/scala/com/github/unchama/seichiassist/subsystems/sharedinventory/System.scala
@@ -1,4 +1,4 @@
-package com.github.unchama.seichiassist.subsystems.shareinventory
+package com.github.unchama.seichiassist.subsystems.sharedinventory
 
 import cats.effect.concurrent.Ref
 import cats.effect.{ConcurrentEffect, SyncEffect}
@@ -7,11 +7,11 @@ import com.github.unchama.datarepository.bukkit.player.BukkitRepositoryControls
 import com.github.unchama.generic.ContextCoercion
 import com.github.unchama.minecraft.bukkit.algebra.BukkitPlayerHasUuid.instance
 import com.github.unchama.seichiassist.meta.subsystem.Subsystem
-import com.github.unchama.seichiassist.subsystems.shareinventory.application.repository.SharedInventoryRepositoryDefinition
-import com.github.unchama.seichiassist.subsystems.shareinventory.bukkit.command.ShareInventoryCommand
-import com.github.unchama.seichiassist.subsystems.shareinventory.domain.SharedFlag
-import com.github.unchama.seichiassist.subsystems.shareinventory.domain.bukkit.InventoryContents
-import com.github.unchama.seichiassist.subsystems.shareinventory.infrastracture.JdbcSharedInventoryPersistence
+import com.github.unchama.seichiassist.subsystems.sharedinventory.application.repository.SharedInventoryRepositoryDefinition
+import com.github.unchama.seichiassist.subsystems.sharedinventory.bukkit.command.ShareInventoryCommand
+import com.github.unchama.seichiassist.subsystems.sharedinventory.domain.SharedFlag
+import com.github.unchama.seichiassist.subsystems.sharedinventory.domain.bukkit.InventoryContents
+import com.github.unchama.seichiassist.subsystems.sharedinventory.infrastracture.JdbcSharedInventoryPersistence
 import org.bukkit.command.TabExecutor
 import org.bukkit.entity.Player
