+++ src/main/scala/com/github/unchama/seichiassist/subsystems/sharedinventory/application/repository/SharedInventoryRepositoryDefinition.scala
@@ -1,4 +1,4 @@
-package com.github.unchama.seichiassist.subsystems.shareinventory.application.repository
+package com.github.unchama.seichiassist.subsystems.sharedinventory.application.repository
 
 import cats.Monad
 import cats.effect.Sync
@@ -6,7 +6,7 @@ import cats.effect.concurrent.Ref
 import com.github.unchama.datarepository.definitions.RefDictBackedRepositoryDefinition
 import com.github.unchama.datarepository.template.RepositoryDefinition
 import com.github.unchama.minecraft.algebra.HasUuid
-import com.github.unchama.seichiassist.subsystems.shareinventory.domain.{
+import com.github.unchama.seichiassist.subsystems.sharedinventory.domain.{
   SharedFlag,
   SharedInventoryPersistence
 }
